Move files around to prevent circular dependencies
