+++ packages/agents-manager/src/hooks/use-should-use-unified-agent.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-restricted-imports */
-import { useUnifiedAiChat } from '../data/use-unified-ai-chat';
+import { useUnifiedAiChat } from './use-unified-ai-chat';
 
 export const useShouldUseUnifiedAgent = () => {
 	const { data: isEligibleViaAPI } = useUnifiedAiChat();
