+++ packages/help-center/tsconfig.json
@@ -14,6 +14,7 @@
 		{ "path": "../viewport" },
 		{ "path": "../odie-client" },
 		{ "path": "../wpcom-checkout" },
+		{ "path": "../agents-manager" },
 		{ "path": "../zendesk-client" },
 		{ "path": "../agents-manager" },
 		{ "path": "../support-articles" }
